client/rpc: rename package name to match rpc and edit migration story
