+++ client/rpc/object.go
@@ -1,4 +1,4 @@
-package httpapi
+package rpc
 
 import (
 	"bytes"
