+++ client/rpc/response.go
@@ -1,4 +1,4 @@
-package httpapi
+package rpc
 
 import (
 	"encoding/json"
