+++ docs/changelogs/v0.21.md
@@ -66,7 +66,8 @@ updated the CoreAPI with new Kubo features but forgot to port thoses to the
 http-client, making it impossible to use them together with the same coreapi
 version.
 
-**TODO(@Jorropo)**: add link to `boxo-migrate` once support for rewriting this import path has been added
+For smooth transition `v0.7.0` of `go-ipfs-http-client` provides updated stubs
+for Kubo `v0.21`.
 
 ### 📝 Changelog
